perf: reduce registration request count #issue-24
